DEV: BROKEN; initial draft of wpi calculation. models are wrong, 5 day weather retrieval from results db is missing
