+++ utils/database_utils/results/wpi_utils.go
@@ -1,4 +1,4 @@
-package weather_pleasantry
+package main 
 
 // Simple math functions and checks go here
 
@@ -54,4 +54,22 @@ func weatherCondPleasantness(cond string, config WeatherPleasantnessConfig) floa
 		return 0
 	}
 	return pleasantness
+}
+
+
+// filterDataByDayRange filters the weather data for a specific range of days
+func filterDataByDayRange(weeklyData []model.WeatherData, startDay, endDay time.Weekday) map[time.Weekday][]model.WeatherData {
+	dailyData := make(map[time.Weekday][]model.WeatherData)
+	for _, data := range weeklyData {
+		timestamp := time.Unix(data.Dt, 0)
+		day := timestamp.Weekday()
+		hour := timestamp.Hour()
+
+		if timeutils.ShouldIncludeDay(day, startDay, endDay) {
+			if hour >= 9 && hour <= 21 { // Include data points between 9 am and 9 pm
+				dailyData[day] = append(dailyData[day], data)
+			}
+		}
+	}
+	return dailyData
 }
